+++ src/excella.hpp
@@ -10,7 +10,7 @@ constexpr const char* version = "Excella v1.12.0";
 extern std::string activeDbPath;
 extern std::string cacheFilePath;
 extern std::vector<std::string> cachedDbPaths;
-extern size_t startHash;
+extern u64 startHash;
 extern int actualTotalEntries;
 extern bool dirty;
